Parameterizable Transformations

Transformations that can contain parameters, e.g. linear transformations or neural networks. The respective transformation also needs to be differentiable with respect to both input and parameters in order to properly optimize the boosting model and additional parameters.

All transformations should be applicable using both the boosting output and the actual input.

I.e. let $f(\cdot)$ denote the function corresponding to a Gradient Boosting model, $\mathbf{x}$ the vector of inputs and $g$ a ParameterizableTransform. Then, $g$ should be applicable as

\[g(f(\mathbf{x}),\mathbf{x};\theta)\]

Where $\theta$ denotes a respective parameter vector of the transformation itself.

This increases flexibility and allows to, for example, build varying coefficient models where the varying coefficients are modelled by individual gradient boosting models, $f_1,...,f_n$:

\[g(f_1(\mathbf{x}),...,f_n(\mathbf{x}),\mathbf{x};\theta)=\alpha_0 + f_1(\mathbf{x})\cdot x_1 + \cdots + f_n(\mathbf{x})\cdot x_n,\]

where $\theta\equiv [\alpha]$.

AdvancedBoosting.ParameterizableTransform — Type

The abstract supertype corresponding to parameterizable transformations.

AdvancedBoosting.IdentityTransform — Type

Applies the identity transformation,

\[g(f(\mathbf{x}), \mathbf{x})=f(\mathbf{x})_{[i_1,i_2,...,i_m]},\]

where $[i_1,i_2,...,i_m]$ denotes the indices defined in target_idx

using AdvancedBoosting
transform = IdentityTransform([1])
transform([2.], zeros(3))

# output

1-element Vector{Float64}:
 2.0

using AdvancedBoosting
transform = IdentityTransform([1])
transform([2.])

# output

1-element Vector{Float64}:
 2.0

AdvancedBoosting.SoftplusTransform — Type

Applies the softplus transform,

\[g(f(\mathbf{x}),\mathbf{x})=\log(\exp(f(\mathbf{x})_{[i_1,i_2,...,i_m]})+1),\]

where $[i_1,i_2,...,i_m]$ denotes the elements defined in target_idx.

using AdvancedBoosting
transform = SoftplusTransform([1])
transform([0.], zeros(3))

# output

1-element Vector{Float64}:
 0.6931471805599453

using AdvancedBoosting
transform = SoftplusTransform([1])
transform([0.])

# output

1-element Vector{Float64}:
 0.6931471805599453

AdvancedBoosting.SigmoidTransform — Type

Applies the sigmoid transform,

\[g(f(\mathbf{x}),\mathbf{x})=\frac{1}{1+\exp(-f(\mathbf{x})_{[i_1,i_2,...,i_m]})},\]

where $[i_1,i_2,...,i_m]$ denotes the elements defined in target_idx.

using AdvancedBoosting
transform = SigmoidTransform([1])
transform([0.], zeros(3))

# output

1-element Vector{Float64}:
 0.5

using AdvancedBoosting
transform = SigmoidTransform([1])
transform([0.])

# output

1-element Vector{Float64}:
 0.5

AdvancedBoosting.MultiTransform — Type

Applies multiple ParameterizableTransforms on the same input and concatenates their outputs. I.e. let $g_1(\cdot,\cdot),...,g_n(\cdot,\cdot)$ denote a set of $n$ ParameterizableTransforms. Then, the $MultiTransform$ computes as

\[g(f(\mathbf{x}),\mathbf{x})=\left(g_1(f(\mathbf{x}),\mathbf{x}),...,g_n(f(\mathbf{x}),\mathbf{x})\right)^T,\]

where the outputs of each individual transform are flattened.

using AdvancedBoosting
transform1 = IdentityTransform([1])
transform2 = SigmoidTransform([2])
transform = MultiTransform([transform1, transform2])

transform([0.,0.], zeros(3))

# output

2-element Vector{Float64}:
 0.0
 0.5

using AdvancedBoosting
transform1 = IdentityTransform([1])
transform2 = SigmoidTransform([2])
transform = MultiTransform([transform1, transform2])

transform([0.,0.])

# output

2-element Vector{Float64}:
 0.0
 0.5

AdvancedBoosting.VaryingCoefficientTransform — Type

Applies the outputs of multiple gradient boosting models as a varying coefficient model. I.e., for an input vector $\mathbf{x}\in\mathbb{R}^M$ and $M$ stacked boosting models,

\[f(\cdot)=\left(f_1(\cdot),...,f_M(\cdot)\right)^T,\]

we have

\[g(f(\mathbf{x}),\mathbf{x};\alpha)=\alpha + \left(f_{(i_1)}(\mathbf{x}),...,f__{(i_K)}(\mathbf{x})\right)^T \mathbf{x}_{(i_1,...,i_K)},\]

where $\alpha$ denotes the intercept of the varying coefficient model and $i_1,...,i_K$ denote the indices defined in target_dims

using AdvancedBoosting
transform = VaryingCoefficientTransform([1,2])

transform([1.,2.], [1.,2.])

# output

1-element Vector{Float64}:
 6.0

AdvancedBoosting.ComposedTransform — Type

Composes two ParameterizableTransforms as follows:

Let $g1,g2$ denote two ParameterizableTransforms, $\mathbf{x}$ some input vector and $\mathbf{f}(\mathbf{x})$ the outputs of one or more boosting models, then

\[g(\mathbf{f}(\mathbf{x}),\mathbf{x})=(g_2\circ g_1)(\mathbf{f}(\mathbf{x}),\mathbf{x})=g_2(g_1(\mathbf{f}(\mathbf{x}),\mathbf{x}),\mathbf{x})\]

Example:

using AdvancedBoosting
g1 = VaryingCoefficientTransform([1,2])
g2 = SoftplusTransform([1])

transform = g2∘g1

transform([1.,2.], [1.,2.])

# output

1-element Vector{Float64}:
 6.00247568513773

AdvancedBoosting.SumTransform — Type

Sums two ParameterizableTransforms as follows:

Let $g1,g2$ denote two ParameterizableTransforms, $\mathbf{x}$ some input vector and $\mathbf{f}(\mathbf{x})$ the outputs of one or more boosting models, then

\[g(\mathbf{f}(\mathbf{x}),\mathbf{x})=(g_1 + g_2)(\mathbf{f}(\mathbf{x}),\mathbf{x})=g_1(\mathbf{f}(\mathbf{x}),\mathbf{x}) + g_2(\mathbf{f}(\mathbf{x}),\mathbf{x})\]

Example:

using AdvancedBoosting
g1 = SoftplusTransform([1])
g2 = SoftplusTransform([1])

transform = g2+g1

transform([1.], [1.])

# output

1-element Vector{Float64}:
 2.6265233750364456