Root Boosting Model

The underlying Gradient Boosting model that can be stacked to create new, more sophisticated models.

You typically don't want to work with these low-level models directly but embed them into higher level models.

AdvancedBoosting.RootBoostingModel — Type

Single boosting model whose inputs, outputs and parameters are of type T<:AbstractFloat.

Can be constructed either via

RootBoostingModel{T}(max_depth::Int64, n_trees::Int64)::RootBoostingModel{T} where T<:AbstractFloat

or

RootBoostingModel(max_depth::Int64, n_trees::Int64)::RootBoostingModel{Float64}

where max_depth defines the maximum depth of each tree (typically, max_depth=1 suffices in a Gradient Boosting model) and n_trees defines the number of trees to include in the model, i.e. the number refinement iterations

Once a is_trained(RootBoostingModel{T})==true, i.e. the model has been fitted to some data, we can give it an input matrix X::Matrix{T} to produce some predictions:

model = RootBoostingModel(1,100)
...(training the model here)
X = randn(5,10) #5 datapoints with 10 features each
predictions = model(X)

It is also possible to use an input vector X::Vector{T}. In that case, X will be treated as a single-row matrix, i.e. an input matrix with a single data point.

model = RootBoostingModel(1,100)
...(training the model here)
X = randn(10) # a single datapoint with 10 features
predictions = model(X)

Notice that the RootBoostingModel deliberately has no API to be trained directly. Rather, a higher level model typically employs several RootBoostingModels and the training happens via the higher level model APIs.

As a helper function, it is also possible to make predictions from a vector of RootBoostingModels as follows:

models = [RootBoostingModel(1,100), RootBoostingModel(1,100)]
...(training the models here)
X = randn(10) # a single datapoint with 10 features
predictions = models(X) #returns a vector of prediction vectors

AdvancedBoosting.is_trained — Function

is_trained(m::RootBoostingModel)::Bool

Check whether a boosting model has already been fitted/trained.